pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "powerschool-adapter"
version = "1.0.1"
description = "A Python adapter for PowerSchool API integration"
readme = "README.md"
requires-python = ">=3.7"
license = { text = "MIT" }
keywords = ["powerschool", "education", "api", "adapter"]
authors = [
    { name = "TONYLABS TECH CO., LTD.", email = "tony.wang@tonylabs.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Education",
]

dependencies = [
    "requests>=2.20.0",
    "diskcache>=5.6.0",
]

[project.optional-dependencies]
dev = [
    "python-dotenv>=1.0.1",
    "faker>=33.3.1"
]

[project.urls]
Homepage = "https://github.com/tonylabs/python-powerschool-adapter"
Repository = "https://github.com/tonylabs/python-powerschool-adapter.git"
"Bug Tracker" = "https://github.com/tonylabs/python-powerschool-adapter/issues"

[tool.black]
line-length = 88
target-version = ["py37"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3